Uhhh a lot of changes
Pushhhhhhhhhhh

+++ VR TEST/Assets/1/Cubes/CubeHealth.cs
@@ -33,7 +33,7 @@ public class CubeHealth : MonoBehaviour
         }
     }
 
-    public void EnemyDamage(int EnemyDamgedNum)
+    public void EnemyDamage(float EnemyDamgedNum)
     {
         cubeHealth -= EnemyDamgedNum;
     }
